Misc Updates.
Textpanel gameobject added to canvas, and PlayerTurnText gameobject added to it to contain the value.
Added a visor to the capsule, and created Prefab
Second Unit Added, added material overrides Red and Blue to help tell the two apart.
Player1 and Player2 are the tags.  Applied said tags to the PlayerPrefabs