Turn Manager v.1
please note i need this script to be put into the editor into a game object in order for it to work. only one of these turn managers should be available at all times. This is untested, and needs to get a run through to ensure functionality.

description of functionality in this version:
1. all units added to turn manager
2. all units sorted by turn counter
3. unit starts its turn
4. character moves (only thing available right now)
5. end character turn
6. loop back to step 2